requirements/local.txt
-r _base.txt

django-debug-toolbar==4.2.0